--- examplesgrpc/function/udfproj/.python-version
@@ -1,1 +1,0 @@
-3.9.12
