--- examplesgrpc/function/udfproj/pyproject.toml
@@ -1,15 +1,0 @@
-[tool.poetry]
-name = "test-python-udf"
-version = "0.1.0"
-description = ""
-authors = ["Numaflow developers"]
-
-[tool.poetry.dependencies]
-python = "~3.9"
-pynumaflow = "0.1.0"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
