+++ pynumaflow/_constants.py
@@ -4,3 +4,4 @@ NUMAFLOW_UDF_CONTENT_TYPE = "NUMAFLOW_UDF_CONTENT_TYPE"
 NUMAFLOW_MESSAGE_KEY = "x-numa-message-key"
 FUNCTION_SOCK_PATH = "/var/run/numaflow/udf.sock"
 SINK_SOCK_PATH = "/var/run/numaflow/udsink.sock"
+DATUM_KEY = "x-numaflow-datum-key"
