--- pynumaflowgrpc/_constants.py
@@ -1,7 +1,0 @@
-APPLICATION_JSON = "application/json"
-APPLICATION_MSG_PACK = "application/msgpack"
-NUMAFLOW_UDF_CONTENT_TYPE = "NUMAFLOW_UDF_CONTENT_TYPE"
-NUMAFLOW_MESSAGE_KEY = "x-numa-message-key"
-FUNCTION_SOCK_PATH = "/var/run/numaflow/udf.sock"
-SINK_SOCK_PATH = "/var/run/numaflow/udsink.sock"
-DATUM_KEY = "x-numaflow-datum-key"
