--- pynumaflowgrpc/function/__init__.py
@@ -1,4 +1,0 @@
-from pynumaflowgrpc.function._dtypes import Message, Messages
-from pynumaflowgrpc.function.handler import HTTPHandler
-
-__all__ = ["Message", "Messages", "HTTPHandler"]
