--- pynumaflowgrpc/sink/__init__.py
@@ -1,4 +1,0 @@
-from pynumaflowgrpc.sink.handler import HTTPSinkHandler
-from pynumaflowgrpc.sink._dtypes import Message, Response, Responses
-
-__all__ = ["HTTPSinkHandler", "Message", "Response", "Responses"]
